Even if a charm is already uploaded, we need the charmhub origin.
AddCharm expects that if a charm hasn't been installed before, the id
and hash of the charm origin are not set.  If the values are not set, we
call refresh with an install action.  If the values are set, call
refresh with a refresh action.  Then we have the case of a previously
uploaded charm with a new name.  The charm has been uploaded already,
but to refresh it later, we need the id and hash to save in the
application.  Ensure that data is provided here.