Handle architecture and series in find
The following brings over architecture and series filtering that has
been bestowed on info. The filtering is done on the client, in the
future this should be done on the server by the API, unfortunately the
API doesn't provide this filtering.

The changes are rather simple and mirror info. It would be nice to
generalise the filtering, but considering the vastly different types
it's actually quite hard to do so.