Readability over conciseness
The following updates a function to export exactly what we mean to
prevent the confusion over what slice was for what.

Returning `[]slice, []slice` looses meaning, adding return types helps
in some cases, but once you export the function as an interface you have
to hope the interface also has the return types! To prevent living with
hope, lets just be explicit!