Standardise commands
The following standardises all the commands in regards to the
architecture and series. Each command then doesn't have to repeat the
same series heavy lifting that the base command does.

The client doesn't require any changes, just the command that filters
based on what is returned. We can generalise the filtering such that we
can have predicates that take in a channel-map and return if it should
be filtered or not.

The code around filtering still needs testing, but marking as a commit
to make it easier to reason about.