Funnel through architecture
The following is the very start of funnelling through architecture. It
makes some big changes by forcing Origin to contain a platform. This
platform now ensures that we pass an architecture, os and series. As
this is just the beginning there are areas that are missing gaps:

 1. Bundle support (charms don't have constraints)
 2. Model migration

Once those two are solved (model migration should be easy enough) then
we can wrap this up.

One problem we have though is that charmhub just fails to download a
charm even if we're sending data for architecture and os and series.