Restrict setting arch constraint
The aim here is to restrict an application architecture constraint once
it's been set. If the architecture is already set and it's the one being
passed in, we treat it as a no-op and move along. Removing the
application removes the constraint so can be re-applied the same as
other contstraints.

The driver behind this is to prevent the deploying of a charm with a
constraint and changing the architecture under the charm with a
different constraint. This will cause a world of pain to try and
reconcile that correctly, re-downloading charms and their resources
along with any profiles that need applying... So the fix is to prevent
it and then if they have done the wrong thing, remove the application
and start again.

Architecture is becoming more pervasive so we need to ensure that we
don't allow foot guns in the codebase.