Derive application arch constraint
Add a strategy for adding adding default application architecture
constraint.

This is just an idea and I'm unsure how successful it is, but
essentially the idea is to check all instances for their own arch and if
it has some the fill it in.

If we have a heterogenous set of machines then use the model to see if
that has one, otherwise use the default architecture.