Fixes an issue with CAData when importing k8s
When we import a cluster with CAData file in the config we fill out the
k8s config struct with the raw CAData as well. This causes an error to
propogate from the k8s client-go code. This change stops that from
happening by using tmp variables to pass the data around.

LP: 1905321