Wire up the client in the refresher with new API
The referesher client can now check for a charm placement using the new
API client.

We only do this in the charmhub side of things as that improves our
single responsiblitly side of things. CharmStore charms don't need to
check for placement. Charmstore charms following the analogy of:

    a charm is a charm is a charm

This is no longer a case as architecture and series dictate that a charm
is a matrix of architecture, series and a charm. So that falls flat.

The nice thing here is that we created a SOLID type that makes it easier
to decide on what gets called depending on what type is used. AMAZING!