Derive default architecture
The following change inserts a derived architecture from either the
model constraints and if one isn't set from the default architecture.
It's important to understand that from now on we fill in the
architecture for all applications and this has consequences for the
distribution of units onto machines. It is now expected that a machine
will be provisioned for a constraint to match the machines instance
hardware characteristic. If non is found a new machine will be brought
up to match that. Although the behavour of this shouldn't have that much
fallout, it is important to note this is a behavioural change.

The code is simple, derive it when adding an application, the fallout is
that the tests need to be updated in lots of places.

We need this to ensure that when deploying a charm the constraints of an
application can't change to ensure that we have the right charm blob for
the machine architecture. This prevents errors down the line in the
status field of the application that an error occured trying to install
the charm.