Potential fixes to bundle deployments
There is an odity where when we fill in bundle changes we end up
creating new machines and I'm not entirely sure this is correct.
Although it doesn't exhibit that problem in manual testing as far as I
can see, it does in testing setups.

I'm a bit wary of this change because we're starting to apply empty
constraints for zero reasons and it seems like we'll need changes to
bundlehandlers to ensure that we're not doing it for no reason.