Parse charm platform
The following introduces the parsing of platform so that importing and
exporting works as expected for the description package. Originally the
description package expected `os/series/arch`, but in reality it should
be `arch/os/series`, with arch being the non-optional value making it
consistent between the lack of optional values.

The rules for platform are the following:

 1. Architecture is mandatory.
 2. OS is optional and can be dropped. Series is mandatory if OS wants
 to be displayed.
 3. Series is also optional.

To indicate something is missing `unknown` can be used in place. Using
Normalize can clean up the output of the Platform output to ensure
consistency.

Examples of this would be:

 1. `<arch>/<os>/<series>`
 2. `<arch>`
 3. `<arch>/<series>`
 4. `<arch>/unknown/<series>`

The code is rather simple as it just splits over a string and plucks the
right parts based on the amount visible.