Resolve origin fields
When attempting to resolve an origin we should populate all the fields
of the origin so that we can correctly send them to and from the server.
These items where missed because of the changes to architecture, but
have now been correctly filled in.

Deploying a local charmhub charm is now possible thanks to this.