Updates to AddPendingResources.
Create a factory which based on charm source, returns an object to
correctly get resource information for the AddPendingResources call into
the backend.

Renamed some variables to reduce confusion between store as in
charmstore, vs store as in the backend state.

Further refinements on how charmhub determines the best list of
resources is needed.