Make UpdateLatestRevisions controller facade work with charmhub
Currently it only pulls new charm revision info from the charmstore;
make it pull from the new charmhub API as well (using Refresh). We
partition the list of the model's charms into charmstore vs charmhub
so we can call each API in bulk. Most of the new charmhub-specific
code is in the new charmhub.go file.

Change RefreshEntity.CreatedAt to a time.Time instead of a string. May
want to do this for other *At fields in there as well?

For testing, I've added a wrench charmrevision/shortinterval that
is checked when the controller starts up. If you create
/var/lib/juju/wrench/charmrevision on the controller with contents
"shortinterval" and then restart the controller it'll set the poll
interval to 10 seconds (!) instead of 24 hours.

Other things:
* Added a debug logger to the charmrevision worker so you can see when
  it fires.
* A few drive-by typo fixes.