Update the upgrade step for 2.9 for charm origin
Now that more fields have been added to charm origin the upgrade steps
should also be added to better reflect that.

There is a worrying part that I'm unsure how to solve and that is around
people whom are already trying out RC1 and RC2 of 2.9. Offically this
won't re-run for them, but is that a bad thing?

 - The items that actually are mandatory are source, revision, channel
 and id. These where already in the version 1 of the upgrade.
 - Platform although important will fallback to the logic to find out
 what it should be if it's missing.
 - Type isn't required for charms and is a requirement for bundles,
 considering we don't store bundles this should be fine.

If what I say holds we should be ok.