Charmhub find summary wordwrap
This is a very simple word wrapping of a line. We know that the summary
has been chopped to one line via the one line function, meaning we can
just look for spaces and hyphens to workout when to wordwrap a line.
With these constraints/restrictions is makes is easy for us to clean up
the find table.

This won't handle pathological cases like long words, but it'll do for
now. It's much easier on the eye than the previous dumpster fire that
find was!