Columnise find to reduce output width
The output find is rather large, now that we have more and more
populated fields, we're potentially showing to much information. Instead
of just removing the fields that look unimportant, we should tailor it
depending on the users needs.

This PR introduces columns for the tabular field. Using --columns flag
with the column alias seems like a decent approach. Other canonical
tools also do this, so it isn't something new (LXD).

With decent defaults it actually looks rather good as well.