Save charm origin platform to state
The issue was that platform wasn't being correctly marshalled to the
state charm origin. This fixes that so that we correctly send the state
platform.

This is potentially one fix that could be lurking in deploying code.
Unfortunately the client does some many things that attemptint to fix
everything is actually quite hard.