Ensure that assess_upgrade_series test can run against develop branch
The develop branch deprecates --machine for juju run and breaks this
test. As a workaround, this commit changes the test to use juju ssh when
issuing reboot requests.

In addition, a 5 second delay is injected between the second (post
series upgrade) reboot and the check for the series reported by the
agent. This delay should be enough to allow the agent to report the
series and invalidate the cached status.