Charmhub info optional flag
Charmhub info now requires a "optional" flag to get the metadata for the
given channel. This is required atm as we use it to use the computed
series.

We currently error out if we can't find it, but I'm starting to think we
should just use the channel map for the same series/revision/arch to
compute that ourselves. That way we don't ever end up in a messed up
state where Juju requests the info but we can't use it because the
stupid metadata isn't there!