Remove state from common client
The common client constructor was bringing in state. We REALLY do NOT
want to include state in the apiserver. It's a slow march forward and
adding more state to the apiserver is counter productive.

This cleans this up and moves the metadata headers to charmhub directly
as we get this for free for others whom are using the client.