Progress bar test susceptible to timing failures
The progress bar tests are susceptible to timing failures if running on
slow CI boxes. The fix is to pass in a clock to prevent that from
happening, so when we run the tests we can pass in our own mock clock
which we can target exactly the time we want to test against.