Resource unit tests
The following adds test for the charmhub repository resolving. It
currently doesn't work as we're waiting for file path to be put back
into the API after they removed it.

Also cleaned up the charmhub headers so that we get ALL the headers for
all scenarios. Previously it only looked in the context, but the
platform could actually be in the actions as well.

It probably should be another branch, but life it too short!