Arm64 unit test fixes with app constraint
The following adds a new method to the connsuite* to ensure that when we
create a testing application that is has the correct arch constraints.
The fact that the constraints wasn't set would mean that it would
default to amd64. This is obviously not what we want and so forcing it
to be the host arch when adding the application solves this.