Bundle YAML for bundles
Charmhub API now exports bundle yaml for bundles rather than bulk
loading them into metadata yaml. This is a good fix as it means bundles
are not an after thought and it promotes bundles to first class
entities.

The change is simple it just expects the bundle-yaml json tag to be set
along with the filters and passes them through the info. The charm
facade can them check the type before performing the unmarshalling of
the data correctly.