Do not allow "all" as valid CharmOrigin values.
Info and other responses in CharmHub have been returning "all" for
series, os and architecture for various charm, and bundles.  "all" is
unknown by juju.  In the Info case, "all" for series or os can be
changed to an empty string.  It's resolved later.  For architecture,
"all" becomes the DefaultArchitecture: amd64.