Ensure we normalize the channel
Although generally not required for speaking to the charmhub API we
should ensure that we present a normalized channel for the request. The
change is really simple and I'm starting to question if we shouldn't
just normalize everywhere, except that you have to really know that
"latest/stable" == "stable". "latest" is a special keyword that is
dropped from a track and it makes it very confusing if you're not aware
of that.