Re-work resource selection
In order to preserve the same functionality of selecting a resource we
ask the server for all the resources for a channel, revision and charm
revision.

This ensures we try our best effort to find the resource we desire from
the API. It's the best we can do with what we've got.

We can then attempt to find the newest resource or we fallback to the
exact same revision.