Fix client facade not setting CanUpgradeTo for charmhub charms
There was an 'if charmURL.Schema == "cs" { ... }' in the code, which
was originally intended to mean "if this charm is not local", but
obviously it prevents charmhub "ch" URLs from being included. Simply
update it to include both stores.

This prevented "juju status --format=yaml" showing the can-upgrade-to
information when new revisions were available.

Prior to this change the following code on
apiserver/facades/client/client/status.go:1190 didn't execute, because
the latestCharm wasn't set.

	if latestCharm, ok := context.allAppsUnitsCharmBindings.latestCharms[*applicationCharm.URL().WithRevision(-1)]; ok && latestCharm != nil {
		if latestCharm.Revision() > applicationCharm.URL().Revision {
			processedStatus.CanUpgradeTo = latestCharm.String()
		}
	}