Add architecture to deploy stack
The following adds architecture to bootstrap when using acceptance
tests. I'm unsure that this covers ALL the tests as it's a bit of a
black box, but it did seem to work correctly when attempting to deploy
an amd64 box with the following args:

```
./deploy_job.py --series=xenial --arch=amd64
```

The amount of indirection in here is NUTS. We probably want to clean
this up or gut it for charmhub.