Sanitize charm origin
The following aims to sanitize the charm origin before it leaves the API
server. Although not required, it ensures that we prevent the leaking of
the platform types out to the client.

The idea is that if we send "arch=amd64" and they send back "arch=all",
then we can replace "all" with "amd64" and the client is none the wiser.

The one difference is that we will recalculate the OS from the series if
the series in the platform is labeled as "all".

This will once and for all kill off the idea of all leaking out.