Ensure we pass arch to boot_context
The boot context takes a arch argument. The arch argument was added to
allow deploying via the deploy_stack function to work. Unfortunately
it's virtually impossible to get all the acceptance tests to run in one
click, so we have to rely on jenkins to tell us when they're failing.