Set CIDR field when instancepoller fakes NICs from instance addresses
When the environment does not support the Networking interface, the
instancepoller worker iterates the addresses reported as part of the
instance info and creates fake NICs based on the discovered addresses.

This commit ensures that the CIDR for the generated NICs gets populated
using the network details associated with each instance address to
prevent the code in networkingcommon from emitting validation errors
when it encounters a blank CIDR value.