Implement identity discovery for ECS tasks
ECS injects the URI to the ECS metadata service in the environment of
each task. The k8sagent scans the environment variables for the presence
of the ECS metadata URI. If found, the agent queries it and populates
the appropriate identity details for the task instance that the k8sagent
is executing under.