Fix oplogSuite.TestStops by inserting row before starting tailer
TestStops was intermittently failing on arm64. I was able to fire up an
arm64 EC2 instance and repro it there, every 50 runs or so. I'm still
not sure why this fixes it, but with this change I can't reproduce it
over thousands of runs. Plus, inserting the row before starting the
tailer is consistent with what TestHonoursInitialTs does.

I also added a cleanup task to drop the fake oplog collection at the
end of the test. This doesn't actually seem to do anything, but I added
it because initially I thought the failure was related to
TestHonoursInitialTs not cleaning up after itself, and TestStops failing
as a result of some interaction. So the AddCleanup here doesn't actually
fix anything, but still seems like the right thing to do.