Ensure we always have an architecture
The following ensures that the client always has an architecture to send
to the controller when deploying a charm or bundle. In theory if none is
sent to the controller we can derive it then, but it's best practice to
at least have one to hand.

The last stage of this is to ensure that we always have one on the
controller so all bases are covered.