Validate incoming origin
To ensure we don't ever end up with a params origin that might not have
enough data for us to process a charm we need to validate the params
origin. The params origin should always have a source and an
architecture. Everything else we can deduce from the defaults or other
things.