Ensure we check for the channel via track/risk
The problem with checking for the channel name was that we may very
occasionally have a branch in there. If that's the case we may not be
able to verify the correct channel is used.

The correct way here is to ensure we check just the track and the risk
and be done with it.

Added unit tests to support this.