Resolve the series and arch when deploying a bundle.
Pathalogical bundles where we use the same charm with a different series
and architecture now correctly install. There is a slight annoyance in
the fact that when attempting to resolve a charm we end up with a series
and architecture from the resolution that we don't want till later. In
that case we empty the series, architecture will always be correct and
then refill it in again at a later stage when we're adding the charm.