Ensure we set the URL to match the origin
Now that URLs can contain series information, they should correctly
represent the same as the origin. The origin will always be correct, but
the URL should now be correct.

This does this correctly, although the code is no longer pure, we have
to check for charmhub schema, as we can no longer assume that the same
channels and platform is correct.