Be less strict when exporting platforms
The following ensures that we're less strict when exporting platforms
for charms. It makes it much easier to integrate with older controllers
if we don't expect them to know everything. Instead we're going to say
say that if you don't know the architecture try and work it out.
Otherwise fill it in with the default architecture.

This fixes bug #1911800 where we couldn't migrate a model correctly.