Use juju rsa key when remoting via ssh and refactor log collection
The log collection logic uses SSHRemote (from remote.py) to change the
permissions on the logs it wants to collect followed by an scp command
to grab the logs.

However, when SSHRemote is *not* using `juju ssh` and instead shells out
to ssh, it was not passing the correct identity file which caused all
ssh commands to fail.

In addition, the log collection code assumed a fixed location for all
logs whereas the location (and available commands) are different
depending on the series used. The log collection code has been
refactored to locate the logs and ensure that they are correctly pulled
from the remote hosts. More specifically:

- When juju-db  is installed from a snap, the logs must be copied to a
temp folder (chmod on /var/snap/... so we can pull the logs does not
seem to work). On xenial to bionic, juju configures mongod to dump logs
to syslog (which is grabbed by default). Finally, as a fallback, if a
log file is found in /var/log/mongodb it will also be pulled.
- The lxd log location also depends on whether it is installed via apt
or via snap. In the latter case, we copy it to a temp folder just as we
do with the snap logs.
- If netplan is detected, we pull /etc/netplan/*.yaml. Otherwise, we
grab /etc/network/interfaces.
- If ifconfig is detected, we pull its output. Otherwise, we grab the
'ip a' output.