Rename charmhub model config value.
It was decided that charmhub shouldn't be hyphenated. It was originally
like this, but was changed to follow the charm-store standardisation.

The code essentially updates the new model-config value (charmhub-url)
from the default server URL or from what was previously set. This is so
we don't break existing model configs that users might have previously
set to a new value.