Expose API extra
With the new changes to the refresh API and the ability to request a
charm or bundle with missing platforms, the following is required to
know where to look next.

The following aims to allow the look up partial data, with the aim to
suggest ways to fix the API request or show the user the potential
solution to their request.