Allow partial refresh requests.
The following allows for partial refresh requests. These refresh
requests will be supported by the charmhub API in that if you have a
partial platform for OS and Series, passing NA for those will return an
error with suggestions about what are available. We can pass those to
the user when resolving the charm, but for now we won't expose them in
this PR.

The idea with the code is that we don't expose the NA to outside the
library, instead if a series or OS is empty, we then swap it out. This
ensures purity of the client API.