Refactor container_networking test
The original test run "echo $X | nc $IP $PORT" over ssh. Since nc blocks
until it sees an EOF on STDIN (e.g. CTRL+D in an interactive session),
the ssh call would never return so the test would get stuck and
eventually timeout. Even if this was a non-blocking call, the test makes
a call to 'more' over ssh to grab the contents that were sent to the nc
listening side.

To avoid blocking while waiting for these calls, this commit:
1) Changes the 'echo | nc' bit to include '-q 0' flag which allows nc to
exit.
2) Changes the 'more' call to a 'tail | sed' invocation that waits for
the data to appear, tees the content and then exits. Having the tail
instead of 'cat' avoids a race between the 'echo | nc' ssh call and the
ssh call for checking the data recorded by the listening end.