Use refresh for resolving charm channels
This introduces the concept of resolving charms with the preferred
channel via the refresh API. This removes the idea of using the info API
to get the possible charm or bundle revisions. Currently we don't
support bundles as the API doesn't support bundles.

The code is actually quite simple and we can drop the charmhub selector
code that was required to source bundle download URLs.

The only difficulty now is ensuring that we provide all the right
updates to the charm URL and origin with the limited data that we have.