Initial implementation
The following is the initial implementation of using charmhub refresh
api instead of using the info one. There are some real gotchas in here,
but essentially this should work for when the series is correct. When
the series isn't correct we will have to fully inspect the error code to
see if it can also help try and determine the correct series.