Update network-health charm to grab interface details via the ip command
The original version used ifconfig and assumed that all regexes would
match. However, the ifconfig output has changed in the newer net-tools
package thus causing the REs to fail and trigger a python runtime
exception as we attempted to use an undefined variable.

The new charm version has been rewritten to use the 'ip' command and
has been tested on trusty, xenial, bionic and focal. In addition, the
actions now use the appropriate shebang to ensure that the charmhelpers
package is visible in the import path.

The updated charm has been released at cs:~juju-qa:network-health-14.