Update forces series tests
Now we're no longer forcing the series of the charm to match the series
of the machine, these tests are changed.

The real problematic one is the TestDeployCharmsEndpointNotImplemented
because it expects some parts to have a series and others to not. To
prevent change the default behaviour everywhere, I've lifted the calls
into the test itself so we can be exact on what we're matching.