Expose entity type
Exposing entity type now allows us to query the refresh API for bundles.
The exposing of the type will then allow us to drop info requests in
download.

This is quite a simple change for the code...

Additionally with recent code changes we can now drop RefreshPlatform
and reuse the same Platform everywhere.