Handle resolving bundles
The resolution of bundles is now handled in the same logic as charms.
The API decides how things should be correctly installed. It uses the
same logic as finding a charm.

Interestingly because charms don't really have a platform it's possible
to do a one hit refresh and it works first time. So it's quicker to
locate a bundle than a charm.