Expose the type as a type from charmhub API
The type (charm or bundle) would be better off typed all the way through
the stack. At the moment it was a type alias, so essentially just a
string at runtime. This forces the type to be an actual type and this
allows us to match on it, which is much cleaner.