Only save charm IDs during FindDownloadURL.
A charm ID in the origin is the only way we have to know if a charm has
been deployed already.  If a charm has not been deployed already, we use
the Install action of Refresh.  If we're trying to upgrade a charm, we
know the ID and use the Refresh action of Refresh.

Using the charm ID on charm upgrades is important as the charm's name is
subject to change.

Add a tests with comments to ensure this behavior continues into the
future.