Always normalize the channel for refresh
We should always normalize the channel passed into the refresh config.
This is because "latest" is a special keyword and that it can and will
be omitted when doing matches.

The code just cleans up how we construct the params conversion, so that
we do it once. It leads to better readability, but more verbose code. I
think this is fine for now.

There is still work to do around picking the right method for
refreshing, but that can come later.