Fix panic in ec2 instances
This was a fun one to locate. There are two issues here:

 1. Using a slice with holes always leads to bad expectations
 2. Expecting an interface to be backed by one instance seems bad

The following PR ensures that we collect the zones from both amzInstance
and sdkInstance. As we seem to be transitioning between libraries we
need to ensure that we correctly handle those locations.

There aren't any unit tests in here so I did add some to at least cover
us for those scenarios.

----

I wish there was a better lint guard around this as it's very bad
pratice to just convert the type without checking.