Clean up instance avil zone names return type
The return type of InstanceAvailabilityZoneNames had a few gotchas in
it.

 1. It was expected that every zone name was in the same position as the
    instance id. Essentially this allowed for typed holes within the
    returning slice.
 2. Each provider had to implement the same logic and that had/has the
    potential to drift over time when developing.

This seems like the wrong type to return for what was expected. Instead
the following commit changes it to be a map that's keyed by the
instance.Id.

Using a map means it's more logical to understand that the zone name
will fill in the value of the map using the instance.id as the key.