Fixes permission for model operator.
A bug was introduced in the 2.9 branch where local run juju binaries in
a kube cluster default to using the service account attached to their
pod. This change wasn't ported to the model operator so it was using a
service account without any permissions.

This change give the model operator it's own service account with the
need permissions to run.