Ensure relation units
The following fixes an issue when a subordinate hasn't been correctly
related on all keys then it would commit suicide. This was caused when
attempting to introduce the error checking linter into juju code base
was incorrectly done (by me!). The fix is simple, only return on an
error not on all errors or nils.

The REALLY interesting part of this piece of work is that if you return
nil from relationsChanged and you do not iterate over all the keys then
all subordinates will commit suicide.

There are no unit tests for this code, it is one big mudball of
integrations from state and internal state. The only way I caught this
was inspecting a full integration test and digging deeper.

The architecture of this code is hard to understand and is built with
ductape, we should revisit at some point.