Export bundle with charm channel
Charmhub relies on channels as a way to identify the correct way to
install a given charm, we need to ensure that this is correctly exported
with the bundle.

Interestingly channel was never outputted for charmstore or local
charms, which could have been a cause of potential bugs when to
replicate an existings deployment based on revision alone.