Export charmhub charm app name
When deploying charmhub charms, only the application name should be used
to identify the charm (this is actually wrong as well, but more on that
later), so the additional information that's with in the URL is only
used to identify a charm and nothing more.

The fix here is that we correctly output just the charm application name
so you can identify it correctly.

---

The real issue is that the application name could have changed since the
charm was originally deployed and instead we should allow charmhub
identifiers to be used instead. The problem I have with this, is that it
becomes very hard to identify what the charm is, we're also leaking out
charmhub implementation details, but that might be a unnecessary evil?