Bundle diff output with charm url names
The following allows the diff'ing of bundles when using charmhub charm
urls. The internal representation of charmhub urls are different from
what users of juju see. So prefix, arch, series charm urls should not be
shown and therefor should not be used for diff'ing exported bundles.

The code is actually simple, use the charmurl name for charmhub charms
and just update to the latest library changes. With the addition of a
few tests, the following should be possible:

    $ juju diff-bundle <(juju export-bundle)
    {}