Output channel in status
The following makes it possible to see the channel of the charm when
deploying a charm. This not only makes it good for charmhub charms, but
should really be done for charmstore charms. Although the channels are
different it hightlights there could be an issue if you're dependant on
a given charm being from a given channel.

The code itself is really simple, just push the channel through the
apiserver and let it land in the status.

Originally revision wasn't shown for charmhub charms, but ongoing
discussion states that it probably should.

.