Use non-greedy wildcard regex to extract uptime value in acceptance tests
The original regex was too greedy and would fail to match the uptime
value thus causing get_uptime to return 0. As a result, the
assess_container_networking test assumed that the controller machine had
not rebooted yet and tried to block until the machine's ssh port became
unavailble (which never did thus causing the test to timeout).