Ensure we normalize the charm url in bundle handler
The resolve charms and endpoints was doing it's job by resolving the
charm urls, except that for charmhub we actually only need the name and
the schema from the resolution.

This is done this way so that we can handle potential renames from the
API at the resolving code (redis to memcache behind the scenes).

There is more work to be done here as channels aren't handled correctly,
but that's do able via changes to bundlechanges repo.